Split C++ into multiple files
- Includes into single header
- Each algorithm variant into its own
  file
