+++ src/clustRviz.h
@@ -1,0 +1,5 @@
+#define ARMA_USE_SUPERLU 0
+
+#include <RcppArmadillo.h>
+#include <RcppEigen.h>
+#define ARMA_64BIT_WORD
